Patched the Signal Order Bug
The signals coming out of the ADC FIFOs were being processed in the
wrong order.  I patched it.  The problem was that I was processing all
the rising signals then all the falling signals.  This didn't work
because I was I was processing several read cycles with each clock of
the processor.